@@ -3,11 +3,12 @@ package com.udacity.notepad.recycler;
 import android.content.Context;
 import android.os.Handler;
 import android.os.Looper;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.udacity.notepad.R;
 import com.udacity.notepad.data.DataStore;
